backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
# Pydantic for data validation and settings management
pydantic>=2.5.0
pydantic-settings>=2.1.0
# Use psycopg2 instead of psycopg2-binary for better Python 3.13 compatibility
psycopg2>=2.9.9
asyncpg>=0.29.0
sqlalchemy>=2.0.24
alembic==1.12.1
pytest==7.4.0
httpx==0.27.0
# AI Service Dependencies
openai==1.3.0
google-generativeai==0.3.0
huggingface-hub==0.16.4
# transformers==4.33.0  # Optional for local models
# torch==2.0.1         # Optional for local models
requests==2.31.0
# Use a more recent version of aiohttp for Python 3.13 compatibility
aiohttp>=3.9.0

# Utilities
tenacity==8.2.2
backoff==2.2.1
python-multipart==0.0.6

# Google Calendar API Dependencies
google-auth==2.23.4
google-auth-httplib2==0.1.0
google-api-python-client==2.104.0

# JWT Authentication
python-jose[cryptography]==3.3.0

# Password hashing
passlib==1.7.4
bcrypt==4.3.0

# Cryptography for API key encryption
cryptography>=41.0.0

# SEO Analysis APIs
google-search-results==2.4.2  # SerpAPI client
beautifulsoup4==4.12.2

# Speech-to-Text Integration
assemblyai==0.43.1

# Twilio Integration for SMS and Voice
twilio==9.0.0
paypalrestsdk==1.13.1

# Dependency Injection
dependency-injector>=4.48.1
aiosqlite>=0.19.0
redis>=5.0.0